Rename converters file
