+++ src/converters.ts
@@ -1,3 +1,4 @@
+import { HTMLElementContent } from './types';
 
 export const kebabToCamel = (string: string): string => string.replace(/-([a-z])/g, (_, item): string => item.toUpperCase());
